Rename struct and split APIManager+curriculum
into APIManager+SemesterCurriculum and APIManager+SemesterCurriculumList
